+++ core/src/main/java/com/bakdata/dedupe/classifier/ClassificationResult.java
@@ -44,7 +44,7 @@ public class ClassificationResult {
      * <p>The general recommendation is to have confidence in [0; 1], but it should not be confused with a proper
      * mathematical probability.</p>
      */
-    float confidence;
+    double confidence;
     /**
      * Additional explanation for humans, such as the similarity and threshold (0.933 >= 0.9) or the name of a rule that
      * triggered.
